Create README.md
EXERCISES: THE EARLY BIRD GETS THE ORM!

You should write three sections. The first should describe the purpose of the app. The second should describe the current state of the app. The third and final section should describe the future improvements you want to make to the app including your notes about the Person class.